requirements.txt
chex==0.1.7
cycler==0.11.0
dm_haiku==0.0.10
einops==0.7.0
jax==0.4.14
matplotlib==3.7.1
numpy==1.23.5
open_spiel==1.3
optax==0.1.7
pandas==1.5.3
PyYAML==6.0.1
seaborn==0.13.0
tbparse==0.0.7
torch==2.0.1+cu118
tqdm==4.66.1